Add function for installing a compilation environment
This is mostly so that regardless of what functions I can contrive
to put into the runtime environment, a compiler won't say stupid
stuff like "unknown function EVAL".

Arguably all the macros should be stuffed into the compilation
environment as well, but at the moment I don't need that so I'm
putting off thinking about it.