Reject nonstandard host types
This fixes a bit of a hole where extrinsicl's typep could refer to
host types, like so:

(deftype foo () 'integer)
([extrinsicl typep] whatever 'foo)

This would end up in the default clause and so be passed to the
host typep without comment. Given how the type system is, this was
a hole allowing host ACE.

It would be more comprehensive to use ctype or something, but I
don't want to add another complex dependency if it's not necessary.