Don't resolve standard atomic type specifiers into classes
Bit of an internal mess here. With the cross-clasp project, I have
to define "classes" that are accessible via find-class but which
are not host classes. So if they're passed to host functions those
host functions will balk. I've previously been thinking that it's
best to use find-class by default, in case the environment is
doing some weird renaming, but it's just too inconvenient. Maybe
I can come up with a more nuanced solution if I ever actually do
have a system that makes STRING name the bit vector class
or whatever else.